Artistv4/test refactor (#85)
* Adds solidity-coverage

* Artist tests refactor complete

* Changeset

* readme

* Adds better ASCII logo

* Makes ticketNumber start from zero
