+++ .changeset/tame-jars-appear.md
@@ -1,0 +1,8 @@
+---
+'@soundxyz/protocol': minor
+---
+
+Refactors Artist.sol tests
+
+- Creates Config class to make test dependencies clearer (no global scope)
+- Breaks tests out into separate files
