+++ protocol/test/helpers.ts
@@ -108,13 +108,7 @@ export const deployArtistProxy = async (artistAccount: SignerWithAddress, soundO
 };
 
 // shifts edition id to the left by 128 bits and adds the token id in the bottom bits
-export const getTokenId = (editionId: number | string, numSold: number | string, quantity?: number) => {
+export const getTokenId = (editionId: number | string, numSold: number | string) => {
   const shiftFactor = BigNumber.from(1).mul(2).pow(128);
-  let tokenId = BigNumber.from(editionId).mul(shiftFactor).add(numSold);
-
-  if (quantity) {
-    tokenId = tokenId.add(quantity + 1);
-  }
-
-  return tokenId;
+  return BigNumber.from(editionId).mul(shiftFactor).add(numSold);
 };
